+++ .gitignore
@@ -1,6 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 .idea
+notes.txt
 
 # dependencies
 /node_modules
@@ -12,6 +13,8 @@
 
 # production
 /build
+.firebaserc
+firebase.json
 
 # misc
 .DS_Store
