+++ src/components/CodingArea/index.js
@@ -1,5 +1,4 @@
-import React from "react";
-import "../css/CodingArea.css";
+import "./index.css";
 
 const CodingArea = ({
   cursorIndex,
