+++ src/components/InfoPanel/index.css
