--- src/css/App.css
@@ -1,52 +1,0 @@
-.App {
-  display: flex;
-
-  text-align: center;
-}
-
-section.codeSample {
-  display: flex;
-  overflow: hidden;
-  align-items: center;
-  flex-direction: column;
-  justify-content: center;
-
-  width: 100%;
-  min-height: 100vh;
-
-  color: white;
-}
-
-section.codeSample h2 {
-  font-size: 2.1rem;
-  font-weight: normal;
-
-  margin: 5px 0 15px;
-
-  text-align: center;
-}
-
-section.codeSample h2 div.mainCategory {
-  font-size: 1.1rem;
-
-  margin: 0 10px;
-
-  color: #6a7487;
-}
-
-.App-logo {
-  height: 13vmin;
-  margin: 0 0 20px 0;
-
-  animation: App-logo-spin infinite 50s linear;
-  pointer-events: none;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
